refactor: simpilfy stats formatter
