+++ packages/core/modern.config.ts
@@ -6,7 +6,7 @@ export default defineConfig({
   buildConfig: [
     {
       ...baseBuildConfig.buildConfig,
-      input: ['src', '!src/client/hmr', '!src/client/overlay.ts'],
+      input: ['src', '!src/client/hmr.ts', '!src/client/overlay.ts'],
     },
     {
       buildType: 'bundle',
@@ -14,7 +14,7 @@ export default defineConfig({
       target: 'es2017',
       dts: false,
       input: {
-        hmr: 'src/client/hmr/index.ts',
+        hmr: 'src/client/hmr.ts',
         overlay: 'src/client/overlay.ts',
       },
       externals: ['./hmr'],
